Messages are now typed with good and bad. Add new Zenko Power
Power feedback messages are now typed according to whether they are good or bad.
Add a new Zenko Spell Celestial Recovery. Expends a large amount of Celestial Energy to instantly recover a sizeable chunk of HP.
Kitsune Magic now scales in power with the number of tails the Kitsune has.